Rename tapis handler file.
